build: add ts config for cli
