+++ tsconfig.cli.json
@@ -1,0 +1,26 @@
+{
+    "compilerOptions": {
+        "target": "es6",
+        "outDir": "dist/cli",
+        "noUnusedLocals": true,
+        "noUnusedParameters": true,
+        "experimentalDecorators": true,
+        "removeComments": true,
+        "noImplicitAny": false,
+        "moduleResolution": "node",
+        "allowSyntheticDefaultImports": true,
+        "lib": [
+            "es6"
+        ],
+        "types": [
+            "node"
+        ]
+    },
+    "include": [
+        "src/**/*",
+        "**/*.d.ts"
+    ],
+    "files": [
+        "src/cli.ts"
+    ]
+}
